+++ sdk/identity/Azure.Identity/tests/CredentialTestBase.cs
@@ -82,15 +82,16 @@ namespace Azure.Identity.Tests
             _listener.EnableEvents(AzureIdentityEventSource.Singleton, EventLevel.Verbose);
 
             var token = Guid.NewGuid().ToString();
-            TransportParams transportParams = new()
+            TransportConfig transportConfig = new()
             {
                 TokenFactory = req => token
             };
-            var factory = MockTokenTransportFactory(transportParams);
+            var factory = MockTokenTransportFactory(transportConfig);
             var mockTransport = new MockTransport(factory);
 
             var config = new CommonCredentialTestConfig()
             {
+                TransportConfig = transportConfig,
                 Transport = mockTransport,
                 TenantId = TenantId,
                 IsUnsafeSupportLoggingEnabled = isSupportLoggingEnabled
@@ -100,12 +101,61 @@ namespace Azure.Identity.Tests
             {
                 Assert.Ignore($"{credential.GetType().Name} is not an MSAL credential.");
             }
-            transportParams.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
+            transportConfig.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
             AccessToken actualToken = await credential.GetTokenAsync(new TokenRequestContext(MockScopes.Default, null), default);
 
             Assert.AreEqual(token, actualToken.Token);
             string expectedPrefix = isSupportLoggingEnabled ? "True" : "False";
             Assert.True(_listener.EventData.Any(d => d.Payload.Any(p => p.ToString().StartsWith($"{expectedPrefix} MSAL"))));
+        }
+
+        [Test]
+        [TestCase(EventLevel.Informational)]
+        [TestCase(EventLevel.Verbose)]
+        public async Task ListenerEventLevelControlsMsalLogLevel(EventLevel eventLevel)
+        {
+            using var _listener = new TestEventListener();
+            _listener.EnableEvents(AzureIdentityEventSource.Singleton, eventLevel);
+
+            var token = Guid.NewGuid().ToString();
+            TransportConfig transportConfig = new()
+            {
+                TokenFactory = req => token
+            };
+            var factory = MockTokenTransportFactory(transportConfig);
+            var mockTransport = new MockTransport(factory);
+
+            var config = new CommonCredentialTestConfig()
+            {
+                TransportConfig = transportConfig,
+                Transport = mockTransport,
+                TenantId = TenantId,
+                IsUnsafeSupportLoggingEnabled = true
+            };
+            var credential = GetTokenCredential(config);
+            if (!CredentialTestHelpers.IsMsalCredential(credential))
+            {
+                Assert.Ignore($"{credential.GetType().Name} is not an MSAL credential.");
+            }
+            transportConfig.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
+            AccessToken actualToken = await credential.GetTokenAsync(new TokenRequestContext(MockScopes.Default, null), default);
+
+            Assert.AreEqual(token, actualToken.Token);
+
+            Assert.True(_listener.EventData.Any(d => d.Level == EventLevel.Informational && d.EventName == "LogMsalInformational"));
+
+            switch (eventLevel)
+            {
+                case EventLevel.Informational:
+                    Assert.False(_listener.EventData.Any(d => d.Level == EventLevel.Verbose && d.EventName == "LogMsalVerbose"));
+                    break;
+                case EventLevel.Verbose:
+                    Assert.True(_listener.EventData.Any(d => d.Level == EventLevel.Verbose && d.EventName == "LogMsalVerbose"));
+                    break;
+                default:
+                    Assert.Fail("Unexpected event level");
+                    break;
+            }
         }
 
         [Test]
@@ -123,25 +173,26 @@ namespace Azure.Identity.Tests
 
             var token = Guid.NewGuid().ToString();
             // Configure the transport
-            TransportParams transportParams = new()
+            TransportConfig transportConfig = new()
             {
                 TokenFactory = req => token
             };
-            transportParams.RequestValidator = req => transportParams.CalledDiscoveryEndpoint |= req.Uri.Path.Contains("discovery/instance");
-            var factory = MockTokenTransportFactory(transportParams);
+            transportConfig.RequestValidator = req => transportConfig.CalledDiscoveryEndpoint |= req.Uri.Path.Contains("discovery/instance");
+            var factory = MockTokenTransportFactory(transportConfig);
             var mockTransport = new MockTransport(factory);
 
             var config = new CommonCredentialTestConfig()
             {
+                TransportConfig = transportConfig,
                 DisableInstanceDiscovery = disable,
                 Transport = mockTransport,
                 TenantId = TenantId,
             };
             var credential = GetTokenCredential(config);
-            transportParams.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
+            transportConfig.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
             AccessToken actualToken = await credential.GetTokenAsync(new TokenRequestContext(MockScopes.Default, null), default);
 
-            Assert.AreNotEqual(disable, transportParams.CalledDiscoveryEndpoint);
+            Assert.AreNotEqual(disable, transportConfig.CalledDiscoveryEndpoint);
             Assert.AreEqual(token, actualToken.Token);
         }
 
@@ -151,16 +202,17 @@ namespace Azure.Identity.Tests
             // Configure the transport
             var token = Guid.NewGuid().ToString();
             string resolvedTenantId = TenantId;
-            TransportParams transportParams = new()
+            TransportConfig transportConfig = new()
             {
                 TokenFactory = req => token
             };
-            var factory = MockTokenTransportFactory(transportParams);
+            var factory = MockTokenTransportFactory(transportConfig);
             var mockTransport = new MockTransport(factory);
 
             var mockResolver = new Mock<TenantIdResolverBase>() { CallBase = true };
             var config = new CommonCredentialTestConfig()
             {
+                TransportConfig = transportConfig,
                 Transport = mockTransport,
                 TenantId = TenantId,
                 RequestContext = new TokenRequestContext(MockScopes.Default, tenantId: Guid.NewGuid().ToString()),
@@ -169,7 +221,7 @@ namespace Azure.Identity.Tests
             };
             var credential = GetTokenCredential(config);
 
-            transportParams.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
+            transportConfig.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
 
             // Assert that Resolver is called and that the resolved tenant is the expected tenant
             mockResolver.Setup(r => r.Resolve(It.IsAny<string>(), It.IsAny<TokenRequestContext>(), It.IsAny<string[]>())).Callback<string, TokenRequestContext, IList<string>>((tenantId, context, additionalTenants) =>
@@ -187,7 +239,7 @@ namespace Azure.Identity.Tests
             var token = Guid.NewGuid().ToString();
             bool observedCae = false;
             bool observedNoCae = false;
-            TransportParams transportParams = new()
+            TransportConfig transportConfig = new()
             {
                 TokenFactory = req => token,
                 RequestValidator = req =>
@@ -220,11 +272,12 @@ namespace Azure.Identity.Tests
                     }
                 }
             };
-            var factory = MockTokenTransportFactory(transportParams);
+            var factory = MockTokenTransportFactory(transportConfig);
             var mockTransport = new MockTransport(factory);
 
             var config = new CommonCredentialTestConfig()
             {
+                TransportConfig = transportConfig,
                 Transport = mockTransport,
                 TenantId = TenantId,
             };
@@ -233,7 +286,7 @@ namespace Azure.Identity.Tests
             {
                 Assert.Ignore("EnableCAE tests do not apply to the non-MSAL credentials.");
             }
-            transportParams.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
+            transportConfig.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
 
             // First call with EnableCae = false
             using (HttpPipeline.CreateClientRequestIdScope("disableCae"))
@@ -269,7 +322,7 @@ namespace Azure.Identity.Tests
             // Configure the transport
             var token = Guid.NewGuid().ToString();
             const string Claims = "myClaims";
-            TransportParams transportParams = new()
+            TransportConfig transportConfig = new()
             {
                 TokenFactory = req => token,
                 RequestValidator = req =>
@@ -297,11 +350,12 @@ namespace Azure.Identity.Tests
                     }
                 }
             };
-            var factory = MockTokenTransportFactory(transportParams);
+            var factory = MockTokenTransportFactory(transportConfig);
             var mockTransport = new MockTransport(factory);
 
             var config = new CommonCredentialTestConfig()
             {
+                TransportConfig = transportConfig,
                 Transport = mockTransport,
                 TenantId = TenantId,
                 RedirectUri = new Uri("http://localhost:8400/")
@@ -311,7 +365,7 @@ namespace Azure.Identity.Tests
             {
                 Assert.Ignore("EnableCAE tests do not apply to the non-MSAL credentials.");
             }
-            transportParams.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
+            transportConfig.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
 
             using (HttpPipeline.CreateClientRequestIdScope("NoClaims"))
             {
@@ -328,6 +382,47 @@ namespace Azure.Identity.Tests
         }
 
         [Test]
+        [NonParallelizable]
+        public async Task TokenContainsRefreshOn()
+        {
+            // Skip test if the credential does not support disabling instance discovery
+            if (!typeof(ISupportsDisableInstanceDiscovery).IsAssignableFrom(typeof(TCredOptions)))
+            {
+                // Assert.Ignore($"{typeof(TCredOptions).Name} does not implement {nameof(ISupportsDisableInstanceDiscovery)}");
+            }
+
+            // Clear instance discovery cache
+            StaticCachesUtilities.ClearStaticMetadataProviderCache();
+
+            var token = Guid.NewGuid().ToString();
+            // Configure the transport
+            TransportConfig transportConfig = new()
+            {
+                TokenFactory = req => token
+            };
+            transportConfig.RequestValidator = req => transportConfig.CalledDiscoveryEndpoint |= req.Uri.Path.Contains("discovery/instance");
+            var factory = MockTokenTransportFactory(transportConfig);
+            var mockTransport = new MockTransport(factory);
+
+            var config = new CommonCredentialTestConfig()
+            {
+                TransportConfig = transportConfig,
+                Transport = mockTransport,
+                TenantId = TenantId,
+            };
+            var credential = GetTokenCredential(config);
+            if (!CredentialTestHelpers.IsMsalCredential(credential))
+            {
+                Assert.Ignore("EnableCAE tests do not apply to the non-MSAL credentials.");
+            }
+            transportConfig.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
+            AccessToken actualToken = await credential.GetTokenAsync(new TokenRequestContext(MockScopes.Default, null), default);
+
+            Assert.AreEqual(token, actualToken.Token);
+            Assert.IsNotNull(actualToken.RefreshOn);
+        }
+
+        [Test]
         public async Task CachingOptionsAreRespected()
         {
             // Skip test if the credential does not support caching options
@@ -336,16 +431,17 @@ namespace Azure.Identity.Tests
                 Assert.Ignore($"{typeof(TCredOptions).Name} does not implement {nameof(ISupportsTokenCachePersistenceOptions)}");
             }
 
-            TransportParams transportParams = new()
+            TransportConfig transportConfig = new()
             {
                 TokenFactory = req => Guid.NewGuid().ToString()
             };
-            var factory = MockTokenTransportFactory(transportParams);
+            var factory = MockTokenTransportFactory(transportConfig);
             var mockTransport = new MockTransport(factory);
             var cache = new MemoryTokenCache();
 
             var config = new CommonCredentialTestConfig()
             {
+                TransportConfig = transportConfig,
                 Transport = mockTransport,
                 TenantId = TenantId,
                 TokenCachePersistenceOptions = cache,
@@ -363,7 +459,7 @@ namespace Azure.Identity.Tests
             {
                 Assert.Ignore($"{credential.GetType().Name} is not an MSAL credential.");
             }
-            transportParams.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
+            transportConfig.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
 
             // Fetch a token to populate the cache
             AccessToken actualToken1 = await credential.GetTokenAsync(new TokenRequestContext(MockScopes.Default, null), default);
@@ -373,6 +469,47 @@ namespace Azure.Identity.Tests
             AccessToken actualToken2 = await credential2.GetTokenAsync(new TokenRequestContext(MockScopes.Default, null), default);
 
             Assert.AreEqual(actualToken1.Token, actualToken2.Token);
+        }
+
+        [Test]
+        public async Task AuthorityHostConfigSupportsdStS()
+        {
+            // Configure the transport
+            var token = Guid.NewGuid().ToString();
+            TransportConfig transportConfig = new()
+            {
+                TokenFactory = req => token,
+                RequestValidator = req =>
+                {
+                    if (req.Uri.Path.EndsWith("/token"))
+                    {
+                        Assert.AreEqual("usnorth-passive-dsts.dsts.core.windows.net", req.Uri.Host);
+                        Assert.AreEqual($"/dstsv2/{TenantId}/oauth2/v2.0/token", req.Uri.Path);
+                    }
+                }
+            };
+            var factory = MockTokenTransportFactory(transportConfig);
+            var mockTransport = new MockTransport(factory);
+
+            var config = new CommonCredentialTestConfig()
+            {
+                TransportConfig = transportConfig,
+                Transport = mockTransport,
+                TenantId = TenantId,
+                AuthorityHost = new("https://usnorth-passive-dsts.dsts.core.windows.net/dstsv2"),
+                RedirectUri = new Uri("http://localhost:8400/")
+            };
+            var credential = GetTokenCredential(config);
+            if (!CredentialTestHelpers.IsMsalCredential(credential))
+            {
+                Assert.Ignore("AuthorityHostConfigSupportsdStS tests do not apply to the non-MSAL credentials.");
+            }
+            transportConfig.IsPubClient = CredentialTestHelpers.IsCredentialTypePubClient(credential);
+
+            // First call to populate the account record for confidential client creds
+            await credential.GetTokenAsync(new TokenRequestContext(MockScopes.Default), default);
+            var actualToken = await credential.GetTokenAsync(new TokenRequestContext(MockScopes.Alternate), default);
+            Assert.AreEqual(token, actualToken.Token);
         }
 
         public class MemoryTokenCache : UnsafeTokenCacheOptions
@@ -464,8 +601,15 @@ namespace Azure.Identity.Tests
                     claimsIsVerified = true;
                     return new ValueTask<AuthenticationResult>(result);
                 });
+
+            var token = Guid.NewGuid().ToString();
+            TransportConfig transportConfig = new()
+            {
+                TokenFactory = req => token,
+            };
             var config = new CommonCredentialTestConfig()
             {
+                TransportConfig = transportConfig,
                 TenantId = TenantId,
                 MockPublicMsalClient = msalPub,
                 MockConfidentialMsalClient = msalConf,
@@ -484,15 +628,16 @@ namespace Azure.Identity.Tests
             Assert.True(claimsIsVerified);
         }
 
-        public class TransportParams
+        public class TransportConfig
         {
             public bool CalledDiscoveryEndpoint { get; set; }
             public bool IsPubClient { get; set; }
             public Action<MockRequest> RequestValidator { get; set; }
             public Func<MockRequest, string> TokenFactory { get; set; }
+            public Action<MockRequest, MockResponse> ResponseHandler { get; set; }
         }
 
-        public static Func<MockRequest, MockResponse> MockTokenTransportFactory(TransportParams transportParams)
+        public static Func<MockRequest, MockResponse> MockTokenTransportFactory(TransportConfig transportConfig)
         {
             return req =>
             {
@@ -505,20 +650,30 @@ namespace Azure.Identity.Tests
                 {
                     response.SetContent(UserrealmResponse);
                 }
-                else
+                else if (req.Uri.Path.Contains("/common/discovery/instance"))
                 {
-                    if (transportParams.IsPubClient || typeof(TCredOptions) == typeof(AuthorizationCodeCredentialOptions) || typeof(TCredOptions) == typeof(OnBehalfOfCredentialOptions))
+                    transportConfig.CalledDiscoveryEndpoint = true;
+                    response.SetContent(CredentialTestHelpers.CreateMockInstanceDiscoveryResponse());
+                }
+                else if (req.Uri.Path.EndsWith("/token"))
+                {
+                    if (transportConfig.IsPubClient || typeof(TCredOptions) == typeof(AuthorizationCodeCredentialOptions) || typeof(TCredOptions) == typeof(OnBehalfOfCredentialOptions))
                     {
-                        response = CredentialTestHelpers.CreateMockMsalTokenResponse(200, transportParams.TokenFactory?.Invoke(req) ?? Guid.NewGuid().ToString(), TenantId, ExpectedUsername, ObjectId);
+                        response = CredentialTestHelpers.CreateMockMsalTokenResponse(200, transportConfig.TokenFactory?.Invoke(req) ?? Guid.NewGuid().ToString(), TenantId, ExpectedUsername, ObjectId);
                     }
                     else
                     {
-                        response.SetContent($"{{\"token_type\": \"Bearer\",\"expires_in\": 9999,\"ext_expires_in\": 9999,\"access_token\": \"{transportParams.TokenFactory?.Invoke(req) ?? Guid.NewGuid().ToString()}\" }}");
+                        response.SetContent($$"""{"token_type": "Bearer","expires_in": 9999,"ext_expires_in": 9999, "refresh_in": 9999,"access_token": "{{transportConfig.TokenFactory?.Invoke(req) ?? Guid.NewGuid().ToString()}}" }""");
                     }
                 }
-                if (transportParams.RequestValidator != null)
+                else if (transportConfig.ResponseHandler != null)
                 {
-                    transportParams.RequestValidator(req);
+                    transportConfig.ResponseHandler(req, response);
+                }
+
+                if (transportConfig.RequestValidator != null)
+                {
+                    transportConfig.RequestValidator(req);
                 }
                 return response;
             };
@@ -702,6 +857,11 @@ namespace Azure.Identity.Tests
             }
         }
 
+        public static Action<object> GetExceptionAction(Exception exceptionToThrow)
+        {
+            return (p) => throw exceptionToThrow;
+        }
+
         public class CommonCredentialTestConfig : TokenCredentialOptions, ISupportsAdditionallyAllowedTenants, ISupportsDisableInstanceDiscovery
         {
             public bool DisableInstanceDiscovery { get; set; }
@@ -711,6 +871,7 @@ namespace Azure.Identity.Tests
             public Uri RedirectUri { get; set; }
             public TokenCachePersistenceOptions TokenCachePersistenceOptions { get; set; }
             public AuthenticationRecord AuthenticationRecord { get; set; }
+            public TransportConfig TransportConfig { get; set; }
             internal TenantIdResolverBase TestTentantIdResolver { get; set; }
             internal MockMsalConfidentialClient MockConfidentialMsalClient { get; set; }
             internal MockMsalPublicClient MockPublicMsalClient { get; set; }