+++ sdk/storage/Azure.Storage.DataMovement.Blobs/src/DataMovementBlobConstants.cs
@@ -7,21 +7,26 @@ namespace Azure.Storage.DataMovement.Blobs
 {
     internal class DataMovementBlobConstants
     {
-        internal class SourceCheckpointData
+        internal class ResourceId
         {
-            internal const int SchemaVersion = 1;
-
-            internal const int VersionIndex = 0;
-            internal const int BlobTypeIndex = VersionIndex + IntSizeInBytes;
-            internal const int DataSize = BlobTypeIndex + OneByte;
+            internal const string BlockBlob = "BlockBlob";
+            internal const string PageBlob = "PageBlob";
+            internal const string AppendBlob = "AppendBlob";
         }
 
-        internal class DestinationCheckpointData
+        internal class SourceCheckpointDetails
         {
-            internal const int SchemaVersion = 2;
+            internal const int DataSize = 0;
+        }
+
+        internal class DestinationCheckpointDetails
+        {
+            internal const int SchemaVersion_3 = 3;
+            internal const int SchemaVersion = SchemaVersion_3;
 
             internal const int VersionIndex = 0;
-            internal const int BlobTypeIndex = VersionIndex + IntSizeInBytes;
+            internal const int PreserveBlobTypeIndex = VersionIndex + IntSizeInBytes;
+            internal const int BlobTypeIndex = PreserveBlobTypeIndex + OneByte;
             internal const int PreserveContentTypeIndex = BlobTypeIndex + OneByte;
             internal const int ContentTypeOffsetIndex = PreserveContentTypeIndex + OneByte;
             internal const int ContentTypeLengthIndex = ContentTypeOffsetIndex + IntSizeInBytes;